SCMOD-7821: Classes renamed to reflect its usage
